pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multiscale-toolbox"
version = "0.2.0"
description = "Reusable multiscale image processing toolbox for Gaussian and Laplacian pyramids in 2D and 3D."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "scipy",
  "matplotlib",
  "pillow",
  "nibabel",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
